(refactor) relocated get_template and render methods in appropriate modules
